# api/docs/swagger.yaml
definitions:
...
  library_service.GenreCreateReq:
    properties:
      name:
        type: string
    type: object
  library_service.GenreGetAllRes:
    properties:
      Genres:
        items:
          $ref: '#/definitions/library_service.GenreRes'
        type: array
    type: object
  library_service.GenreRes:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /genres:
    get:
      consumes:
      - application/json
      description: Get all genre items
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library_service.GenreGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all genre items
      tags:
      - Genre
    post:
      consumes:
      - application/json
      description: Create a new genre item
      parameters:
      - description: Genre data
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/library_service.GenreCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library_service.GenreRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create genre item
      tags:
      - Genre
  /genres/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an genre item by ID
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Genre item deleted
          schema:
            type: string
        "400":
          description: Invalid menu item ID
          schema:
            type: string
        "404":
          description: Menu item not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete menu genre
      tags:
      - Genre
    get:
      consumes:
      - application/json
      description: Get an genre item by ID
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library_service.GenreRes'
        "400":
          description: Invalid menu item ID
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get genre item
      tags:
      - Genre
    put:
      consumes:
      - application/json
      description: Update an existing genre item
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated genre item data
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/library_service.GenreCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/library_service.GenreRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Menu item not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update genre item
      tags:
      - Genre
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"